requirements.txt
pytest==7.4.3
selenium==4.14.0
allure-pytest==2.13.2
requests==2.31.0
pytest-xdist==3.3.1
